[FIXUP] Make use of modelized_list correctly
We currently aren't using `modelized_list` at all, except to add to it
(but not read from it).  This might have been intentional to get the
tests to pass, but it is a major detriment to the code performance that
we would have gained from switching to this model scheme.

This makes it so that the `key` is instead used to define the list,
instead of the method name, since a key can be represented with `:Key`
or `:key`, and define the same method.  This fixes the specs where a key
of both `:TIMESTAMP` and `:timestamp` exist, but hopefully that is more
of a simulation, and the azure API is more ridged about not returning
keys like `:id` and `:ID` on the same endpoint.